RACE CONDITION
A race condition occurs when more than one thread is
performing a series of actions on shared resources, and
several possible outcomes can exist based on the order of the
actions from each thread. The below code is not thread-safe
and the value could be initialized more than once, as checkthen-act
(check for null, then initialize) that lazily initializes
the field is not atomic.