DATA RACE
A data race occurs when 2 or more threads try to access the
same non-final variable without synchronization. Not using
synchronization may lead to making changes which are not
visible to other threads, so reading the stale data is possible,
which in turn may have consequences such as infinite loops,
corrupted data structures, or inaccurate computations. This
code might result in an infinite loop, because the reader thread
may never observe the changes made by the writer threads: