THE SYNCHRONIZED KEYWORD
The synchronized keyword is used to prevent different
threads executing the same code block simultaneously. It
guarantees that since you acquire a lock (by entering the
synchronized block), data, which is protected by this lock, can
be manipulated in exclusive mode, so the operation can be
atomic. Also, it guarantees that other threads will observe the
result of the operation after they acquire the same lock.